requirements.txt
asgiref==3.5.2
crispy-tailwind==0.5.0
Django==4.1.2
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.14.0
mysqlclient==2.1.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
